Move todos handler and router to http package
