+++ http/todo_test.go
@@ -1,4 +1,4 @@
-package todo
+package http_test
 
 import (
 	"io/ioutil"
@@ -6,6 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	todoHttp "github.com/overkilling/overkill-todo-monolith-api/http"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -13,7 +14,7 @@ func TestGetTodosHandler(t *testing.T) {
 	res := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "http://localhost:3000/todos", nil)
 
-	getTodosHandler()(res, req)
+	todoHttp.NewTodosHandler()(res, req)
 
 	content, _ := ioutil.ReadAll(res.Body)
 	assert.Equal(t, "[{\"todo\":\"Some task\"}]", string(content))
